modularize a bit
