+++ packages/bot/src/index.ts
@@ -1,13 +1,17 @@
 /**************************************************************************
  * This file is generated by running 🏕️ scripts/generate_sdk_index.sh 🏕️ *
  **************************************************************************/
+export * from './actions/channel'
+export * from './actions/encryption'
+export * from './actions/interaction'
+export * from './actions/message'
+export * from './actions/moderation'
+export * from './actions/permissions'
+export * from './actions/pinning'
+export * from './actions/tip'
+export * from './actions/types'
 export * from './bot'
-export { default as simpleAppAbi } from '@towns-protocol/generated/dev/abis/ISimpleApp.abi'
-export { getSmartAccountFromUserId } from './getSmartAccountFromUserId'
-export type {
-    BotIdentityConfig,
-    BotIdentityMetadata,
-    ERC8004Endpoint,
-    ERC8004Registration,
-    ERC8004TrustModel,
-} from './identity-types'
+export * from './event-dispatcher'
+export * from './getSmartAccountFromUserId'
+export * from './identity-types'
+export * from './snapshot-getter'
